Improve Id management : when adding an Id in a list, find the lowest but not taken Id before adding the last Id + 1

+++ tools/def.h
@@ -2,5 +2,5 @@
 #include <QPair>
 
 typedef QPair<QString, QString> QStringPair;
-typedef size_t Id;
+typedef quint64 Id;
 #define InvalidId (Id)-1
